requirements.txt
asgiref==3.2.3
Django==3.0.4
image==1.5.28
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
termcolor==1.1.0
